requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
fastapi
uvicorn
httpx
pandas
